fix: moved serializers to server-sided
