+++ src/main/java/com/hamusuke/numguesser/server/game/event/events/CardsOpenEvent.java
@@ -1,8 +1,8 @@
 package com.hamusuke.numguesser.server.game.event.events;
 
-import com.hamusuke.numguesser.game.card.Card;
+import com.hamusuke.numguesser.server.game.card.ServerCard;
 
 import java.util.List;
 
-public record CardsOpenEvent(List<? extends Card> cards) implements GameEvent {
+public record CardsOpenEvent(List<ServerCard> cards) implements GameEvent {
 }
