+++ src/main/java/com/hamusuke/numguesser/server/game/event/events/PlayerStartAttackEvent.java
@@ -1,19 +1,19 @@
 package com.hamusuke.numguesser.server.game.event.events;
 
-import com.hamusuke.numguesser.game.card.Card;
+import com.hamusuke.numguesser.server.game.card.ServerCard;
 import com.hamusuke.numguesser.server.network.ServerPlayer;
 
 public final class PlayerStartAttackEvent extends PlayerEvent {
-    private final Card card;
+    private final ServerCard card;
     private final boolean isCancellable;
 
-    public PlayerStartAttackEvent(ServerPlayer player, Card card, boolean isCancellable) {
+    public PlayerStartAttackEvent(ServerPlayer player, ServerCard card, boolean isCancellable) {
         super(player);
         this.card = card;
         this.isCancellable = isCancellable;
     }
 
-    public Card getCard() {
+    public ServerCard getCard() {
         return this.card;
     }
